More information on Sequoia
Not sure how you want this formatted, but notably, [this](https://twitter.com/_m_vh/status/891164234620338176) suggests to me it's probably version 1.  Also, running `strings` on `WRITEIN.OFF` produces ordered write in votes, and I think a more experienced reverser might be able to get the names of the voters themselves from the file, though I could not.  I included the latter detail but not the former in this (very minor) PR.